# life/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the Life Pratics project",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/pacientes": {
            "get": {
                "tags": [
                    "Admin|Paciente"
                ],
                "summary": "Exibe uma lista de pacientes",
                "operationId": "eec3e4a5bf58cd7743d4b279037e3ebc",
                "responses": {
                    "200": {
                        "description": "Lista de pacientes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin|Paciente"
                ],
                "summary": "Armazena um novo paciente no banco de dados",
                "operationId": "dfe799772046613ad3f7b75167d9756c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pac_nome": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "pac_data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1980-01-01"
                                    },
                                    "pac_email": {
                                        "type": "string",
                                        "example": "joao.silva@example.com"
                                    },
                                    "pac_password": {
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paciente criado com sucesso"
                    }
                }
            }
        },
        "/pacientes/{id}": {
            "get": {
                "tags": [
                    "Admin|Paciente"
                ],
                "summary": "Exibe um paciente específico",
                "operationId": "c9e5621b4983dde500c80317e0b38896",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do paciente"
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin|Paciente"
                ],
                "summary": "Atualiza um paciente existente no banco de dados",
                "operationId": "52720f8c7401f8c7518e6d78547d5b25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pac_nome": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "pac_data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1980-01-01"
                                    },
                                    "pac_email": {
                                        "type": "string",
                                        "example": "joao.silva@example.com"
                                    },
                                    "pac_password": {
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paciente atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin|Paciente"
                ],
                "summary": "Remove um paciente do banco de dados",
                "operationId": "c8adedb54e04d8d21199b810d335c543",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente removido com sucesso"
                    }
                }
            }
        },
        "/pacientes/consultas": {
            "get": {
                "tags": [
                    "Paciente",
                    "Consulta"
                ],
                "summary": "Lista as consultas do paciente autenticado",
                "operationId": "a0a33999f6435b96bb18f398e8cbc88e",
                "responses": {
                    "200": {
                        "description": "Lista de consultas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Paciente",
                    "Consulta"
                ],
                "summary": "Marca uma nova consulta",
                "operationId": "4212dfe0faeabb5df082818887e79795",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "med_codigo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cons_horario": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-11T10:00:00Z"
                                    },
                                    "procedimentos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consulta marcada com sucesso"
                    },
                    "400": {
                        "description": "Erro ao marcar a consulta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pacientes/consultas/{consulta_id}": {
            "get": {
                "tags": [
                    "Paciente",
                    "Consulta"
                ],
                "summary": "Exibe uma consulta específica do paciente autenticado",
                "operationId": "4bb557178c693409d5c97859b7698800",
                "parameters": [
                    {
                        "name": "consulta_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da consulta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cons_codigo": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "cons_horario": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-07-11T10:00:00Z"
                                        },
                                        "cons_particular": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "pac_codigo": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "med_codigo": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "procedimentos": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "proc_codigo": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "proc_nome": {
                                                        "type": "string",
                                                        "example": "Consulta de rotina"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}